Change update policy of ephemeris and ionospheric parameter.
* For synchronization, G_Packet_Ephemeris and G_Packet_Iono_UTC are
added. The space_node in StreamProcessor is updated not by
StreamProcessor, but by NAV. This is because NAV requires to change the
state of space_node to select most appropriate ephemeris when update
with raw information is performed.
* space_node is associated with these packets by a pointer and
G_Packrt_Raw's space node is also changed from a reference to a pointer.