Change RINEX NAV header reading policy; iono and UTC are optional
Previously, ionospheric and UTC parameters were treated as mandatory
header items included in a RINEX NAV file, however correctly, it is
optional ones.
Moreover, a case in which only iono or UTC is included should be
considered.

In addiiton, RINEX.h is fixed to have its classes having template
parameter.